+++ ethereum-types/src/uint.rs
@@ -8,6 +8,7 @@
 
 #[cfg(feature = "codec")]
 use impl_codec::impl_uint_codec;
+#[cfg(feature = "rlp")]
 use impl_rlp::impl_uint_rlp;
 #[cfg(feature = "serialize")]
 use impl_serde::impl_uint_serde;
@@ -19,6 +20,7 @@ construct_uint! {
 	/// Unsigned 64-bit integer.
 	pub struct U64(1);
 }
+#[cfg(feature = "rlp")]
 impl_uint_rlp!(U64, 1);
 #[cfg(feature = "serialize")]
 impl_uint_serde!(U64, 1);