+++ parity-util-mem/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog].
 - Updated dlmalloc to 0.2.1. [#452](https://github.com/paritytech/parity-common/pull/452)
 - Updated dependencies. [#470](https://github.com/paritytech/parity-common/pull/470)
 
+### Breaking
+- Updated `ethereum-types` to 0.10. [#463](https://github.com/paritytech/parity-common/pull/463)
+
 ## [0.7.0] - 2020-06-24
 - Added `const_size` to `MallocSizeOf` to optimize it for flat collections. [#398](https://github.com/paritytech/parity-common/pull/398)
 - Exported `MallocShallowSizeOf`. [#399](https://github.com/paritytech/parity-common/pull/399)